Fully works now from localhost instead of official domain
Proxy most things to the official server and modify request, that need
to be modified. At least config.json is required to be over https, so
there is some dummy https stuff now, that requires a "I accept the risk"
when first visiting the site. Can be swapped an actual cert if necessary